requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
python-dotenv==1.0.0
groq==0.9.0
APScheduler==3.10.4
python-dateutil
google-api-python-client==2.111.0
google-auth-oauthlib==1.2.0
spotipy==2.23.0
